Add custom hook to retrieve the state of the public key
